third_party/blink/web_tests/flag-specific/disable-site-isolation-trials/external/wpt/preload/prefetch-types-expected.txt

Content-Type: text/plain
This is a testharness.js-based test.
FAIL Prefetch as= should work when consumed as  (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as= should work when consumed as image (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as= should work when consumed as script (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as= should work when consumed as style (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as= should work when consumed as document (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=image should work when consumed as  (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=image should work when consumed as image (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=image should work when consumed as script (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=image should work when consumed as style (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=image should work when consumed as document (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=script should work when consumed as  (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=script should work when consumed as image (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=script should work when consumed as script (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=script should work when consumed as style (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=script should work when consumed as document (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=style should work when consumed as  (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=style should work when consumed as image (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=style should work when consumed as script (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=style should work when consumed as style (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=style should work when consumed as document (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=document should work when consumed as  (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=document should work when consumed as image (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=document should work when consumed as script (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=document should work when consumed as style (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
FAIL Prefetch as=document should work when consumed as document (http://web-platform.test:8001) assert_equals: expected (string) "empty" but got (undefined) undefined
Harness: the test ran to completion.

#EOF
#EOF
